Refactor Slideshow and centralize tests
